fix: formatação no padrão brasileiro dos preços de produto recolocados

+++ vSpring/autoagenda/src/main/java/br/com/autoagenda/autoagenda/repositorios/ProdutoRepository.java
@@ -14,5 +14,5 @@ public interface ProdutoRepository extends CrudRepository<Produto, Integer>{
 	long countByEstoqueAtualLessThan(int nivelCondicao);
 	
 	 @Query("SELECT SUM(p.precoCusto) FROM Produto p")
-	 Double sumTotalPrecoCusto(); 
+	 Float sumTotalPrecoCusto(); 
 }
